// frontend/src/models/updateContactDto.ts
/**
 * Generated by orval v8.2.0 🍺
 * Do not edit manually.
 * 企账通SCRM API
 * 企账通SCRM系统API文档
 * OpenAPI spec version: 1.0
 */

export interface UpdateContactDto {
  /** 联系人姓名 */
  name?: string;
  /** 联系电话 */
  phone?: string;
  /** 联系邮箱 */
  email?: string;
  /** 微信号 */
  wechat?: string;
  /** 职位 */
  position?: string;
  /** 部门 */
  department?: string;
  /** 生日 */
  birthdate?: string;
  /** 标签(JSON数组) */
  tags?: string;
  /** 备注 */
  remark?: string;
}